Renamed sagemaker-xgboost-framework to sagemaker-xgboost-container.
