+++ docker/0.82/final/Dockerfile.cpu
@@ -1,13 +1,13 @@
-FROM xgboost-framework-base:0.82-cpu-py3
+FROM xgboost-container-base:0.82-cpu-py3
 
 LABEL com.amazonaws.sagemaker.capabilities.accept-bind-to-port=true
 
 COPY requirements.txt /requirements.txt
 RUN pip install -r /requirements.txt && rm /requirements.txt
 
-COPY dist/sagemaker_xgboost_framework-1.0-py2.py3-none-any.whl /sagemaker_xgboost_framework-1.0-py2.py3-none-any.whl
-RUN pip install --no-cache /sagemaker_xgboost_framework-1.0-py2.py3-none-any.whl && \
-    rm /sagemaker_xgboost_framework-1.0-py2.py3-none-any.whl
+COPY dist/sagemaker_xgboost_container-1.0-py2.py3-none-any.whl /sagemaker_xgboost_container-1.0-py2.py3-none-any.whl
+RUN pip install --no-cache /sagemaker_xgboost_container-1.0-py2.py3-none-any.whl && \
+    rm /sagemaker_xgboost_container-1.0-py2.py3-none-any.whl
 
 
 # Copied from AIApplicationsBaseDockerCPUImage
@@ -40,5 +40,5 @@ ENV SAGEMAKER_HTTP_PORT=8080
 # Make sure that we expose our HTTP port.
 EXPOSE $SAGEMAKER_HTTP_PORT
 
-ENV SAGEMAKER_TRAINING_MODULE sagemaker_xgboost_framework.training:main
-ENV SAGEMAKER_SERVING_MODULE sagemaker_xgboost_framework.serving:main
+ENV SAGEMAKER_TRAINING_MODULE sagemaker_xgboost_container.training:main
+ENV SAGEMAKER_SERVING_MODULE sagemaker_xgboost_container.serving:main
